Some tweaks

+++ Online-Cab-BookingApp/src/main/java/com/cg/sprint/service/impl/CustomerServiceImpl.java
@@ -6,13 +6,16 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.cg.sprint.entity.Admin;
 import com.cg.sprint.entity.Customer;
+import com.cg.sprint.exception.AdminNotFoundException;
+import com.cg.sprint.exception.CustomerNotFoundException;
 import com.cg.sprint.repository.CustomerRepository;
 import com.cg.sprint.service.CustomerService;
 
 @Service
 public class CustomerServiceImpl implements CustomerService {
+
+	String excep = "No such customer found";
 
 	@Autowired
 	CustomerRepository customerRepository;
@@ -26,25 +29,33 @@ public class CustomerServiceImpl implements CustomerService {
 
 	// update customer method
 	@Override
-	public Customer updateCustomer(Customer cust) {
+	public Customer updateCustomer(Customer cust) throws CustomerNotFoundException {
 		Optional<Customer> cusOpt = customerRepository.findById(cust.getCustomerId());
 		Customer cust1 = null;
-		cust1 = cusOpt.get();
-		cust1.setCustomerId(cust.getCustomerId());
-		cust1.setUsername(cust.getUsername());
-		cust1.setPassword(cust.getPassword());
-		cust1.setAddress(cust.getAddress());
-		cust1.setMobileNumber(cust.getMobileNumber());
-		cust1.setEmail(cust.getEmail());
-		customerRepository.save(cust1);
-
-		return cust1;
+		if (cusOpt.isPresent()) {
+			cust1 = cusOpt.get();
+			cust1.setCustomerId(cust.getCustomerId());
+			cust1.setUsername(cust.getUsername());
+			cust1.setPassword(cust.getPassword());
+			cust1.setAddress(cust.getAddress());
+			cust1.setMobileNumber(cust.getMobileNumber());
+			cust1.setEmail(cust.getEmail());
+			customerRepository.save(cust1);
+			return cust1;
+		} else {
+			throw new CustomerNotFoundException(excep);
+		}
 	}
 
 	// delete customer using customer id method
 	@Override
 	public void deleteCustomer(Long customerId) {
-		customerRepository.deleteById(customerId);
+		Optional<Customer> cusOpt = customerRepository.findById(customerId);
+		if (cusOpt.isPresent()) {
+			customerRepository.deleteById(customerId);
+		} else {
+			throw new CustomerNotFoundException(excep);
+		}
 	}
 
 	// view all customers method
@@ -56,27 +67,23 @@ public class CustomerServiceImpl implements CustomerService {
 	// view customer using customer id method
 	@Override
 	public Customer viewCustomer(Long customerId) {
-		return customerRepository.findUserByCustomerId(customerId);
+		Optional<Customer> cusOpt = customerRepository.findById(customerId);
+		if (cusOpt.isPresent()) {
+			return customerRepository.findUserByCustomerId(customerId);
+		} else {
+			throw new CustomerNotFoundException(excep);
+		}
 	}
 
 	// validate customer method
 	@Override
-	public Boolean validateCustomer(Long customerId,String password) {
-		Customer custId = customerRepository.findUserByCustomerId(customerId);
-
-		try {
-			if ( custId.getPassword().equals(password)) {
-
-				return true;
-				
-
-			}else {
-				throw new Exception();
-			}
-		} catch (Exception e) {
-			return false;
+	public String validateCustomer(Long customerId, String password) {
+		Customer cust = customerRepository.findUserByCustomerId(customerId);
+		if (cust!=null&&cust.getPassword().equals(password)) {
+			return "Login Successful";
+		} else {
+			throw new CustomerNotFoundException("Either customer id or password is incorrect");
 		}
-
 	}
 
 }
